+++ diesel/src/expression/count.rs
@@ -9,6 +9,27 @@ use types::BigInt;
 /// As with most bare functions, this is not exported by default. You can import
 /// it specifically as `diesel::expression::count`, or glob import
 /// `diesel::expression::dsl::*`
+///
+/// # Examples
+///
+/// ```rust
+/// # #[macro_use] extern crate diesel;
+/// # include!("src/doctest_setup.rs");
+/// # use diesel::expression::dsl::*;
+/// #
+/// # table! {
+/// #     users {
+/// #         id -> Integer,
+/// #         name -> VarChar,
+/// #     }
+/// # }
+/// #
+/// # fn main() {
+/// #     use self::animals::dsl::*;
+/// #     let connection = establish_connection();
+/// assert_eq!(Ok(1), animals.select(count(name)).first(&connection));
+/// # }
+/// ```
 pub fn count<T: Expression>(t: T) -> Count<T> {
     Count {
         target: t,
@@ -25,7 +46,7 @@ pub fn count<T: Expression>(t: T) -> Count<T> {
 /// it specifically as `diesel::expression::count_star`, or glob import
 /// `diesel::expression::dsl::*`
 ///
-/// # Example
+/// # Examples
 ///
 /// ```rust
 /// # #[macro_use] extern crate diesel;