+++ diesel/src/expression/functions/aggregate_ordering.rs
@@ -55,8 +55,50 @@ macro_rules! ord_function {
 
 ord_function!(max, Max, "MAX",
 "Represents a SQL `MAX` function. This function can only take types which are
-ordered.");
+ordered.
+
+# Examples
+
+```rust
+# #[macro_use] extern crate diesel;
+# include!(\"src/doctest_setup.rs\");
+# use diesel::expression::dsl::*;
+#
+# table! {
+#     users {
+#         id -> Integer,
+#         name -> VarChar,
+#     }
+# }
+#
+# fn main() {
+#     use self::animals::dsl::*;
+#     let connection = establish_connection();
+assert_eq!(Ok(Some(8)), animals.select(max(legs)).first(&connection));
+# }
+");
 
 ord_function!(min, Min, "MIN",
 "Represents a SQL `MIN` function. This function can only take types which are
-ordered.");
+ordered.
+
+# Examples
+
+```rust
+# #[macro_use] extern crate diesel;
+# include!(\"src/doctest_setup.rs\");
+# use diesel::expression::dsl::*;
+#
+# table! {
+#     users {
+#         id -> Integer,
+#         name -> VarChar,
+#     }
+# }
+#
+# fn main() {
+#     use self::animals::dsl::*;
+#     let connection = establish_connection();
+assert_eq!(Ok(Some(4)), animals.select(min(legs)).first(&connection));
+# }
+");