+++ diesel/src/query_source/mod.rs
@@ -7,8 +7,6 @@ pub mod joins;
 use backend::Backend;
 use expression::{Expression, SelectableExpression, NonAggregate};
 use query_builder::*;
-#[doc(hidden)]
-pub use self::joins::{InnerJoinSource, LeftOuterJoinSource};
 use types::{FromSqlRow, HasSqlType};
 
 pub use self::joins::JoinTo;
@@ -49,18 +47,4 @@ pub trait Table: QuerySource + AsQuery + Sized {
 
     fn primary_key(&self) -> Self::PrimaryKey;
     fn all_columns() -> Self::AllColumns;
-
-    fn inner_join<T>(self, other: T) -> InnerJoinSource<Self, T> where
-        T: Table,
-        Self: JoinTo<T, joins::Inner>,
-    {
-        InnerJoinSource::new(self, other)
-    }
-
-    fn left_outer_join<T>(self, other: T) -> LeftOuterJoinSource<Self, T> where
-        T: Table,
-        Self: JoinTo<T, joins::LeftOuter>,
-    {
-        LeftOuterJoinSource::new(self, other)
-    }
 }