+++ migrations/sqlite/20170407152306_add_nullable_table/up.sql
@@ -1,0 +1,4 @@
+CREATE TABLE nullable_table (
+    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
+    value INTEGER
+);